CRP-1428: Globally unique Transcript ID

This MR extends the TranscriptId with the height of the source subnet at which the IDKG protocol is started. We modify the constructor of the Transcript ID to get as input the height. We also make the following changes:
* We introduce an `update_height()` method to the transcript ID to increase the source height. If the height is smaller than `self.source_height` then an error is returned.
* The transcript_id.id is changed from `usize` to `u64`.
* The `context_data` now includes the height of the source subnet. 

See merge request dfinity-lab/public/ic!4096

+++ rs/consensus/src/ecdsa/utils.rs
@@ -495,7 +495,7 @@ pub(crate) mod test_utils {
     }
 
     // Creates a TranscriptID for tests
-    pub(crate) fn create_transcript_id(id: usize) -> IDkgTranscriptId {
+    pub(crate) fn create_transcript_id(id: u64) -> IDkgTranscriptId {
         dummy_idkg_transcript_id_for_tests(id)
     }
 
@@ -595,7 +595,7 @@ pub(crate) mod test_utils {
     // Creates a test signature input
     pub(crate) fn create_sig_inputs_with_height(caller: u8, height: Height) -> TestSigInputs {
         let transcript_id = |offset| {
-            let val = caller as usize;
+            let val = caller as u64;
             create_transcript_id(val * 214365 + offset)
         };
